requirements.txt
astroid==2.9.0
autopep8==1.6.0
backcall==0.2.0
Box2D==2.3.10
cloudpickle==2.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fonttools==4.28.2
future==0.18.2
gym==0.21.0
gym-algorithmic==0.0.1
gym-legacy-toytext==0.0.1
ipykernel==6.5.1
ipython==7.30.0
isort==5.10.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
MDP==3.6
mdptoolbox-hiive==4.0.3.1
nest-asyncio==1.5.1
networkx==2.6.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pycodestyle==2.8.0
pydot==1.4.2
pyglet==1.5.21
Pygments==2.10.0
pylint==2.12.1
pymdptoolbox==4.0b3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.0
wcwidth==0.2.5
wrapt==1.13.3
